Rename Context -> baseOptionsProvider
