+++ src/index.ts
@@ -1,6 +1,6 @@
 import { useCroods } from './useCroods'
 import { useStore } from './useStore'
-import { useBaseOptions, CroodsProvider } from './Context'
+import { useBaseOptions, CroodsProvider } from './baseOptionsProvider'
 import { Fetch } from './Fetch'
 import { useHydrate } from './useHydrate'
